// x-pack/plugins/security_solution/common/api/detection_engine/prebuilt_rules/publish_prebuilt_rules/publish_prebuilt_rules.gen.ts
/*
 * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
 * or more contributor license agreements. Licensed under the Elastic License
 * 2.0; you may not use this file except in compliance with the Elastic License
 * 2.0.
 */

/*
 * NOTICE: Do not edit this file manually.
 * This file is automatically generated by the OpenAPI Generator, @kbn/openapi-generator.
 *
 * info:
 *   title: Publish Prebuilt Rules API endpoint
 *   version: 1
 */

import { z } from 'zod';

import { RuleResponse } from '../../model/rule_schema/rule_schemas.gen';
import { RuleSignatureId, RuleRevision } from '../../model/rule_schema/common_attributes.gen';

export type RuleToPublish = z.infer<typeof RuleToPublish>;
export const RuleToPublish = z.object({
  rule_id: RuleSignatureId,
  revision: RuleRevision,
});

export type PublishPrebuiltRulesRequestBody = z.infer<typeof PublishPrebuiltRulesRequestBody>;
export const PublishPrebuiltRulesRequestBody = z.object({
  /**
   * The list of rules to publish
   */
  rules: z.array(RuleToPublish),
});
export type PublishPrebuiltRulesRequestBodyInput = z.input<typeof PublishPrebuiltRulesRequestBody>;

export type PublishPrebuiltRulesResponse = z.infer<typeof PublishPrebuiltRulesResponse>;
export const PublishPrebuiltRulesResponse = z.object({
  /**
   * The list of published rules
   */
  published_rules: z.array(RuleResponse).optional(),
});